+++ packages/brick_rest_generators/test/rest_model_serdes_generator/test_enum_as_string.dart
