+++ packages/brick_rest_generators/test/rest_model_serdes_generator_test.dart
@@ -1,20 +1,20 @@
 import 'package:analyzer/dart/element/element.dart';
 import 'package:brick_build/generators.dart';
 import 'package:brick_rest/rest.dart';
-import 'package:brick_rest_generators/rest_class_generator.dart';
+import 'package:brick_rest_generators/rest_model_serdes_generator.dart';
 import 'package:test/test.dart';
 import 'package:source_gen/source_gen.dart';
 import 'package:brick_build/testing.dart';
 
-import 'rest_class_generator/test_enum_as_string.dart' as _$enumAsString;
-import 'rest_class_generator/test_ignore_from_to.dart' as _$ignoreFromTo;
+import 'rest_model_serdes_generator/test_enum_as_string.dart' as _$enumAsString;
+import 'rest_model_serdes_generator/test_ignore_from_to.dart' as _$ignoreFromTo;
 
 final _generator = TestGenerator();
-final folder = 'rest_class_generator';
+final folder = 'rest_model_serdes_generator';
 final generateReader = generateLibraryForFolder(folder);
 
 void main() {
-  group('RestClassGenerator', () {
+  group('RestModelSerdesGenerator', () {
     group('@Rest', () {
       test('enumAsString', () async {
         await generateExpectation('enum_as_string', _$enumAsString.output);
@@ -41,7 +41,7 @@ class TestGenerator extends AnnotationSuperGenerator<RestSerializable> {
   /// Given an [element] and an [annotation], scaffold generators
   List<SerdesGenerator> buildGenerators(Element element, ConstantReader annotation) {
     final serializableGenerator =
-        RestClassGenerator(element, annotation, repositoryName: repositoryName);
+        RestModelSerdesGenerator(element, annotation, repositoryName: repositoryName);
     return serializableGenerator.generators;
   }
 }
