+++ packages/brick_sqlite_generators/test/sqlite_model_serdes_generator/test_sqlite_unique.dart
